backend/requirements.txt
fastapi
uvicorn
python-multipart
openai
python-dotenv
faiss-cpu
